chatapp/requirements.txt
chainlit==1.0.504
python-dotenv==1.0.1